committed

+++ src/main/java/com/sunglowsys/repository/HotelInventoryRepository.java
@@ -1,10 +1,11 @@
 package com.sunglowsys.repository;
 
 import com.sunglowsys.domain.HotelInventory;
+import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HotelInventoryRepository extends JpaRepository<HotelInventory, Integer> {
+public interface HotelInventoryRepository extends JpaRepository<HotelInventory, Long> {
 
 }
